+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All notable changes will be documented in this file.
 
+## 5.0.0
+
+### BREAKING
+
+- BREAKING: ESM-only. CommonJS is no longer supported.
+- BREAKING: Drop support for Node.js < v20
+- BREAKING: No default export. Named export `createClient()` should now be used. Same arguments.
+- BREAKING: Drop usage of `node-fetch` - rely on the environments native `fetch` method.
+
+Migrating from v4 or below looks like:
+
+```diff
+-import PicoSanity from 'picosanity'
++import {createClient} from 'picosanity'
+
+-const client = new PicoSanity(options)
++const client = createClient(options)
+```
+
+Using Node.js 18 or older _probably_ still works as long as it has `fetch()` available globally and supports ESM natively, but there are no guarantees for that. Node 20 is now the lowest supported version of Node.js.
+
 ## 4.1.1
 
 ### Fixed